solidrpc: more petstore tests

+++ SolidRpc.Tests/Swagger/CodeGen/UrlParam/Services/IUrlParam.cs
@@ -10,8 +10,8 @@ namespace SolidRpc.Tests.Swagger.CodeGen.UrlParam.Services {
         /// </summary>
         /// <param name="p">The number to proxy</param>
         /// <param name="cancellationToken"></param>
-        Task<long> ProxyIntegerInPath(
-            long p,
+        Task<int> ProxyIntegerInPath(
+            int p,
             CancellationToken cancellationToken = default(CancellationToken));
     
     }
